# Faster --stacktrace:on by avoiding linked list