--- a/koch.nim.cfg
+++ b/koch.nim.cfg
@@ -5,3 +5,11 @@
   # workaround see https://github.com/nim-lang/Nim/pull/14291
   --lib:lib
 @end
+
+@if nimHasCgenVersion1:
+@else:
+  # for bootstrap.
+  # `FR_` etc are no longer there; we could also add more logic to support this
+  # during bootstrap but it may not be worth it
+  --stacktrace:off
+@end