--- a/lib/nimbase.h
+++ b/lib/nimbase.h
@@ -24,6 +24,12 @@ __AVR__
 #ifndef NIMBASE_H
 #define NIMBASE_H
 
+#ifndef NIM_CGEN_VERSION
+  // We increment `NIM_CGEN_VERSION` each time a change is needed, this is
+  // used to make bootstrapping work when changes are needed in this file.
+  #define NIM_CGEN_VERSION 0
+#endif
+
 /*------------ declaring a custom attribute to support using LLVM's Address Sanitizer ------------ */
 
 /*
@@ -87,13 +93,16 @@ __AVR__
 #  define __DECLSPEC_SUPPORTED 1
 #endif
 
+
 /* calling convention mess ----------------------------------------------- */
 #if defined(__GNUC__) || defined(__LCC__) || defined(__POCC__) \
                       || defined(__TINYC__)
   /* these should support C99's inline */
   /* the test for __POCC__ has to come before the test for _MSC_VER,
      because PellesC defines _MSC_VER too. This is brain-dead. */
-#  define N_INLINE(rettype, name) inline rettype name
+//#  define N_INLINE(rettype, name) inline rettype name
+// #  define N_INLINE(rettype, name) inline rettype name
+#  define N_INLINE(rettype, name) __attribute__((always_inline)) rettype name
 #elif defined(__BORLANDC__) || defined(_MSC_VER)
 /* Borland's compiler is really STRANGE here; note that the __fastcall
    keyword cannot be before the return type, but __inline cannot be after
@@ -503,16 +512,24 @@ typedef char* NCSTRING;
 #  endif
 #endif
 
-typedef struct TFrame_ TFrame;
-struct TFrame_ {
-  TFrame* prev;
-  NCSTRING procname;
-  NI line;
-  NCSTRING filename;
-  NI16 len;
-  NI16 calldepth;
-  NI frameMsgLen;
-};
+#if NIM_CGEN_VERSION < 1
+  typedef struct {
+    NCSTRING procname;
+    NI line;
+    NCSTRING filename;
+    NI16 len;
+    NI frameMsgLen;
+  } TFrame;
+#endif
+
+#if NIM_CGEN_VERSION == 1
+  // PRTEMP
+  typedef struct {
+    NCSTRING procname;
+    NI line;
+    NCSTRING filename;
+  } TFrame;
+#endif
 
 #define NIM_POSIX_INIT  __attribute__((constructor))