--- a/lib/nimbase.h
+++ b/lib/nimbase.h
@@ -100,9 +100,8 @@ __AVR__
   /* these should support C99's inline */
   /* the test for __POCC__ has to come before the test for _MSC_VER,
      because PellesC defines _MSC_VER too. This is brain-dead. */
-//#  define N_INLINE(rettype, name) inline rettype name
-// #  define N_INLINE(rettype, name) inline rettype name
-#  define N_INLINE(rettype, name) __attribute__((always_inline)) rettype name
+#  define N_INLINE(rettype, name) inline rettype name
+// #  define N_INLINE(rettype, name) __attribute__((always_inline)) rettype name
 #elif defined(__BORLANDC__) || defined(_MSC_VER)
 /* Borland's compiler is really STRANGE here; note that the __fastcall
    keyword cannot be before the return type, but __inline cannot be after