--- a/lib/nimbase.h
+++ b/lib/nimbase.h
@@ -25,8 +25,8 @@ __AVR__
 #define NIMBASE_H
 
 #ifndef NIM_CGEN_VERSION
-  // We increment `NIM_CGEN_VERSION` each time a change is needed, this is
-  // used to make bootstrapping work when changes are needed in this file.
+  // We increment `NIM_CGEN_VERSION` each time an incompatible change is needed,
+  // this is used to make bootstrapping work.
   #define NIM_CGEN_VERSION 0
 #endif
 
@@ -93,7 +93,6 @@ __AVR__
 #  define __DECLSPEC_SUPPORTED 1
 #endif
 
-
 /* calling convention mess ----------------------------------------------- */
 #if defined(__GNUC__) || defined(__LCC__) || defined(__POCC__) \
                       || defined(__TINYC__)
@@ -101,7 +100,6 @@ __AVR__
   /* the test for __POCC__ has to come before the test for _MSC_VER,
      because PellesC defines _MSC_VER too. This is brain-dead. */
 #  define N_INLINE(rettype, name) inline rettype name
-// #  define N_INLINE(rettype, name) __attribute__((always_inline)) rettype name
 #elif defined(__BORLANDC__) || defined(_MSC_VER)
 /* Borland's compiler is really STRANGE here; note that the __fastcall
    keyword cannot be before the return type, but __inline cannot be after