* this is essentially the 1st part of https://github.com/nim-lang/Nim/pull/13630 which I'm splitting off for simplicity. It turns the linked list into a thread local growable array of TFrame as well as other optimizations
* also works with --stacktracemsgs:on|off; 
* fixes a "multiple evaluation" issue in `stackframes.setFrameMsg`

## example 1
nim r -d:danger --stacktrace:on --stacktracemsgs:off main.nim
before PR: 3.1s
after PR: 2.2s

```nim
when true:
  import times
  var x = 0.uint
  proc fun(i: uint) {.inline.} = x+=i
  proc main() =
    let t = cpuTime()
    let n = 1000_000_000
    for i in 0..<n:
      fun(cast[uint](i))
    echo cpuTime() - t
    doAssert x > 0
    echo x
    proc bar()=echo getStacktrace()
    bar()
  main()
```

## example 2
after compiling nim with `-d:release --stacktrace:on --stacktracemsgs:off`, compiling nim twice; the 2nd time takes 14.9s before PR and 11.7s after PR. I'm measuring the 2nd compilation because the 1st one involves unrelated costs (C compiler) whereas the 2nd one would skip these thanks to `betterRun`

## after this PR
* remove (filename, procname) and push a single location index (as int) which is an index into a lookup table; this will give a further performance boost as well as added flexibility such as getting column info (and even source line info, can be user defined), for free.