@Araq 2020-06-29
Where is the proof that this is faster? As I said in https://github.com/nim-lang/Nim/pull/13630