@timotheecour 2020-06-29
> Where is the proof that this is faster? As I said in #13630

updated top post with 2 examples