@Araq 2020-06-29
Ok, any idea *why* it is faster? also pinging @c-blake for enlightenment. :-)