@timotheecour 2020-06-29
each optimization is cumulative (as you can verify by instrumenting code):
* instead of just `nimln_` (which updates both line+filename) I'm distinguishing when I need to update just line or both filename+line (see nimRefreshFile2 + nimRefreshLine2)

* linked list involve pointer chasing (at variable offset), whereas with array the offset is at a fixed TFrame.sizeof; and I can replace `calldepth` + `prev` by a single index `frameIndex` that indexes into the array. The array accesses are cache friendly.

* after this, the 2nd part (mentioned in top post) gets you an additional performance boost