@Araq 2020-06-29
> linked list involve pointer chasing 

But is there any pointer chasing? The chasing only enters the picture when you iterate over the list to produce a report. Which is not what should be optimized for!