@timotheecour 2020-06-29
> But is there any pointer chasing?

yes, look at `nimFrame`; it must dereference `framePtr` pointer and update that pointer. + all other points mentioned.