@c-blake 2020-06-29
First, in general I am very in favor of an array based stack vs linked list.  The main motive for a linked list, I should think, would be Stackless Python like super dense memory management of stack frames for like a billion green threads.  Density can likely be recovered just with a user-tunable small initial array.  That said, the list is there and it works/is debugged/etc.

As far as the why goes..Free list updates to link in & unlink at the frame allocat/dealloc time are a *little* bit more work than changing the size of an array.  They don't seem like much, but then the effect in question is also not very big.  (3.1-2.2)/1e9\*4GHz = 3.5 CPU clock cycles while a no-op funcall/ret is like 14 and the pipeline is like 12 stages.  So, that microbenchmark is clearly in one "CPU function", and only about 1/4 of a pipeline depth.

Some concrete ways to investigate the "whys" might be on Linux doing `perf stat -d ./testA` and `perf stat -d ./testB` where A&B come from the linked list vs. array versions.  At this kind of pipeline dependency/code layout kind of optimization, PGO can also make a big difference..maybe a bigger difference than the links vs. an array.  https://forum.nim-lang.org/t/6295 explains how in general.

I was trying to do a PGO build of that little billion call test program to report PGO-d vs just LTO-d but while the `-fprofile-generate` and run phases worked, the post `-fprofile-use` build always prints out "No stack trace available" { though it does so in 0.397s vs 1.36 for the profile-generate phase ;-)  Often the profile-generating version runs 3-6x slower than the properly optimized one}.  I can work on this a little more later.  I may have to simplify it and disassemble.

With only a 3.5 cycle delta, it is also ***very*** worth testing both gcc and clang backends and worth testing on both Intel and AMD cpus (and ideally ARM, but that's probably asking too much).  Such a small difference may have ***very*** unstable performance properties across various compilation/deployment contexts.  Conclusion drawing caution is warranted.