@c-blake 2020-06-29
Ok.  With a PGO build of that small test program and `--stacktrace:on --stacktracemsgs:off` which I was neglecting before, I just got "with PGO: 0.985 sec" and with just -O3/flto/etc I got "2.205 sec".  So, 2.24x faster with PGO (all with the linked list, not touching this PR at all), but a much bigger delta than @timotheecour was reporting. That test was on an i7-6700k @4.2GHz with gcc-10.1.

I suspect that if you did PGO builds both with & without the PR the perf delta might vanish.  Also, if you did a PGO for the whole compiler you might get it 2x faster compile times as well which might even be noticeable if you use a `tcc` backend for devel work, but that would rquire some build scripting care (thought it sounds like timothee is halfway there...).

Now, all this may argue both ways in terms of this PR.  Most people do not do PGO builds.  So, maybe this means "awesome - no need to do a PGO build to get this boost!" or, OTOH, it could say "performance sensitive people can already do PGO - don't fix what isn't broken".  It depends on how annoying it is to uses to have/write a benchmark program to collect profile data.  I just have this little `nim-pgo` script I use. But I can say based on my experiment that it's even more worthwhile than my initial guess in this context.