@c-blake 2020-06-29
Oh, and just for the record I used this, not the original program  but I do not believe all that extraneous stuff changes anything { and the when true baffled me :-) } :
```Nim
var x = 0.uint
proc fun(i: uint) {.inline.} = x+=i
proc main() =
  for i in 0 ..< 1_000_000_000: fun(cast[uint](i))
  echo x, getStacktrace()
main()
```