@Araq 2020-06-29
This benchmark seems flawed, you need a recursive test ``fun`` and ensure the optimizer doesn't remove the recursion. Otherwise the debug frame could be moved out of ``fun`` and out of the loop too.