@c-blake 2020-06-29
That's definitely a real hazard.  I did not disassemble both to confirm.