@Araq 2020-06-29
In fact, use https://github.com/Araq/fosdem2020/blob/master/bintrees_manual.nim as the benchmark or something similar.