@c-blake 2020-06-29
For that `bintrees_manual.nim` with a command argument of eleven (`11`, so 2048 trees of depth 4 up to 32 trees of depth 10) to stay close to L1, PGO makes only marginal deltas (order 12% 2.32 ms vs 2.67ms).

Meanwhile `--stacktrace:on --stacktracemsgs:off` makes a full 4x difference (9.10 ms vs 2.32 ms).  So, the stacktrace on/off overhead seems very substantial in this program which might make it a great microbenchmark {good pointer @Araq :-) } for this..maybe 4x slower with linked lists and 3x with an array.  Dunno, but at 4x it does seem like a big overhead maybe worth worrying about, especially if `stacktrace:on` ever becomes on by default in release mode, but I'm guessing you already know that.

I've held off from grabbing his PR and recompiling Nim and all that jazz.  I think he should probably follow up a bit using this `bintrees_manual.nim`.  In his defense, @timotheecour did do the Nim compiling itself which is a big hairy program unlikely to have gcc cheat like you mentioned and he find some noticeable delta (though I would guess still smaller than full PGO on the whole compiler).

I do think it all warrants deeper investigation, but *a priori* "But why?" and "Really?" and "on AMD?  On ARM?" are good questions to keep asking. :-) And some benchmark almost as simple as that first one but uncheatable may still be useful if we really want to know.  Anyway, not much more to add.  Sorry I couldn't offer more enlightenment out of the gate. :-)