@Araq 2020-10-27
- We don't know why it's faster this way.
- We do know that it's more code to maintain.
- The goal of getting better performance for fulldebug builds when we already have support for native zero cost stacktraces seems questionable.